pyproject.toml
[project]
name = "research-radar"
version = "0.1.0"
description = "Research paper discovery and analysis tool."
authors = [
    { name = "Bar Haim", email = "barha@il.ibm.com" },
]
readme = "README.md"
requires-python = ">=3.11.13"
dependencies = [
    "docling>=2.58.0",
    "gradio>=5.49.1",
    "hf-xet>=1.2.0",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-core>=1.0.1",
    "langchain-google-vertexai>=3.0.3",
    "langchain-huggingface>=1.0.0",
    "langchain-ibm>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.2",
    "yt-dlp>=2025.12.8",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pylint>=4.0.2",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
